Added a function-sharing module.
