+++ ts/Series/PackedBubble/PackedBubbleSeries.ts
@@ -49,6 +49,11 @@ const {
         bubble: BubbleSeries
     }
 } = SeriesRegistry;
+import D from '../RAFDeferDataLabels.js';
+const {
+    initDataLabels,
+    initDataLabelsDefer
+} = D;
 import U from '../../Core/Utilities.js';
 const {
     addEvent,
@@ -485,39 +490,6 @@ class PackedBubbleSeries extends BubbleSeries {
     }
 
     /**
-     * PackedBubble needs a custom initDataLabels function
-     * to initiate the animation in a different way
-     * than for other series (drawDataLabels is called each frame)
-     * @private
-     */
-    public initDataLabels(this: PackedBubbleSeries): SVGElement {
-        const series = this,
-            dlOptions = series.options.dataLabels;
-
-        if (!series.dataLabelsGroup) {
-            const dataLabelsGroup = this.initDataLabelsGroup();
-
-            // Apply the dataLabels.style not only to the
-            // individual dataLabels but also to the entire group
-            if (!series.chart.styledMode && dlOptions?.style) {
-                dataLabelsGroup.css(dlOptions.style);
-            }
-
-            // Initialize the opacity of the group to 0 (start of animation)
-            dataLabelsGroup.attr({ opacity: 0 });
-
-            if (series.visible) { // #2597, #3023, #3024
-                dataLabelsGroup.show();
-            }
-
-            return dataLabelsGroup;
-        }
-
-        series.dataLabelsGroup.attr({ opacity: 1 });
-        return series.dataLabelsGroup;
-    }
-
-    /**
      * Packedbubble has two separate collecions of nodes if split, render
      * dataLabels for both sets:
      * @private
@@ -704,21 +676,7 @@ class PackedBubbleSeries extends BubbleSeries {
 
     public init(): PackedBubbleSeries {
         seriesProto.init.apply(this, arguments);
-
-        const dlOptions = this.options.dataLabels;
-
-        // drawDataLabels() fires for the first time after
-        // dataLabels.animation.defer time unless
-        // the dataLabels.animation = false or dataLabels.defer = false
-        // or if the simulation is disabled
-        if (!dlOptions?.defer ||
-            !this.options.layoutAlgorithm?.enableSimulation) {
-            this.deferDataLabels = false;
-        } else {
-            syncTimeout((): void => {
-                this.deferDataLabels = false;
-            }, dlOptions ? animObject(dlOptions.animation).defer : 0);
-        }
+        initDataLabelsDefer.call(this);
 
         /* eslint-disable no-invalid-this */
 
@@ -1299,7 +1257,6 @@ interface PackedBubbleSeries extends DragNodesSeries, NetworkgraphSeries {
     yData: BubbleSeriesType['yData'];
     zData: BubbleSeriesType['zData'];
     zoneAxis: BubbleSeriesType['zoneAxis'];
-    initDataLabels: PackedBubbleSeries['initDataLabels'];
     getPointsCollection(): Array<PackedBubblePoint>;
     indexateNodes: NetworkgraphSeries['indexateNodes'];
     markerAttribs: BubbleSeriesType['markerAttribs'];
@@ -1320,6 +1277,7 @@ extend(PackedBubbleSeries.prototype, {
     pointValKey: 'value',
     requireSorting: false,
     trackerGroups: ['group', 'dataLabelsGroup', 'parentNodesGroup'],
+    initDataLabels: initDataLabels,
     alignDataLabel: seriesProto.alignDataLabel,
     indexateNodes: noop as NetworkgraphSeries['indexateNodes'],
     onMouseDown: DragNodesComposition.onMouseDown,
